vm2-api-server/requirements.txt
fastapi==0.115.0
uvicorn==0.33.0
kafka-python==2.1.5
redis==5.2.1
pydantic==2.10.0 